+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.3.2 (2024-04-29)
+* Fixed Generic Resolver class scanner
+* Added namespace map filter check
+
 ## v0.3.1 (2024-04-26)
 * Fixed namespaceMap accessor